pyproject.toml
[project]
name = "django-crisp-modals"
description = "Django Crispy Forms inside Bootstrap Modals"
authors = [
    {name = "Michel Fodje",email = "michel4j@gmail.com"}
]
readme = "README.md"
license = "MIT"
packages = [
    { include = "crisp_modals", from = "." }
]
requires-python = ">=3.10"
dependencies = [
    "django-crispy-forms>=2.3",
    "django (>=4.2,<6.0.0)",
    "crispy-bootstrap5 (>=2024.10,<2025.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
dynamic = ["version"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry]
version = "0.0.0"

[tool.poetry.requires-plugins]
poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
